requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.41.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0 